refactor ipc into multiple modules
Things were getting way too chaotic. Now we have multiple smaller files,
each roughly concerned with a different problem.
